extract corpse looting
