+++ crates/client/src/app/input/mod.rs
@@ -1,4 +1,5 @@
 pub mod body;
+pub mod corpse_looting;
 pub mod crafting;
 pub mod hotbar_assignment;
 pub mod inventory;
